NtlmMessage.writeSecurityBuffer(): Write out empty data correctly
Current behavior is not to write any data in output buffer if supplied
a zero-length data buffer to write-out.
This is not quite correct though, as we must write that we have no data.
I.e. we should at least write we have a zero-length data which is
layed at offset 0 in the output buffer

NOTE: This will affect both NEGOTIATE_MESSAGE buffer construction
and AUTHENTICATE_MESSAGE buffer construction.